Less Math, more GUI
Instead of using an aditional else to test if x is bigger than y or the
other way around and isntead directly uses Math.Min and Math.max in the
Formula to calculate the random Number.

Actually put the GUI together code wise, don't know how it looks though.
And for some reason I couldn't put the listener on the button so
pressing it does nothing.